L-S-Programming/src/org/howard/edu/lsp/oopfinal/question1/SongsDatabaseTest.java
package org.howard.edu.lsp.oopfinal.question1;

import static org.junit.Assert.*;
import org.junit.*;

import java.util.Set;

public class SongsDatabaseTest {

    @Test
    public void testAddSong() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        Set<String> rapSongs = db.getSongs("Rap");
        assertTrue(rapSongs.contains("Savage"));
    }

    @Test
    public void testGetGenreOfSong() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        assertEquals("Rap", db.getGenreOfSong("Savage"));
    }
    
    @Test
    public void testGetGenreOfNonExistingSong() {
        SongsDatabase db = new SongsDatabase();
        assertNull(db.getGenreOfSong(""));
    }

    @Test
    public void testGetSongs() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        db.addSong("Rap", "Gin and Juice");
        Set<String> rapSongs = db.getSongs("Rap");
        assertEquals(2, rapSongs.size());
        assertTrue(rapSongs.contains("Savage"));
        assertTrue(rapSongs.contains("Gin and Juice"));
    }
}
